create auth.js
