+++ src/data/api.js
@@ -1,11 +1,10 @@
 import token from './token'
 import { throwError } from './error'
 import polyglot from '../i18n'
-import { loggedOut } from './analytics'
 import { reset } from './reset'
 
 // https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API/Using_Fetch#Supplying_request_options
-async function wrappedFetch(url, body) {
+export async function wrappedFetch(url, body) {
   const headers = {
     authorization: token.authzHeader(),
     'content-type': 'application/json',
@@ -59,48 +58,6 @@ export async function gql(query) {
   }
      
   return response.data
-}
-
-export async function login(email, returnTo) {
-  token.reset()
-
-  let loginUrl = `auth/login?client-id=${encodeURIComponent(token.key())}&auth-email=${encodeURIComponent(email)}`
-
-  if (returnTo) {
-    loginUrl += `&return-to=${encodeURIComponent(returnTo)}`
-  }
-
-  // const response = await wrappedFetch(loginUrl)
-  const response = [
-    {
-      name: 'SIL',
-      logoURL: 'https://static.sil.org/brand-tool/glyph1.png',
-      RedirectURL: 'https://example.org/login'
-    },
-    {
-      name: 'Google',
-      logoURL: 'https://www.naschenweng.info/wp-content/uploads/2010/10/google-logo.png',
-      RedirectURL: 'https://example.org/login'
-    },
-  ]
-
-  if (response.length === 1) {
-    window.location = response[0].RedirectURL
-  } else {
-    // store these auth providers
-    // push to another view
-    // render the options
-    // use the RedirectURL for the window.location in that view
-    // destroy the storage
-  }
-}
-
-export function logout() {
-  window.location = `${process.env.BASE_API_URL}/auth/logout?token=${encodeURIComponent(token.pair())}`
-
-  clearLocalData()
-
-  loggedOut()
 }
 
 export async function upload(formData) {
